requirements.txt
playwright==1.50.0
python-dotenv==1.1.0
